修改ratio float

+++ ch-user-core/src/main/java/com/ai/ch/user/dao/mapper/bo/ShopInfo.java
@@ -13,7 +13,7 @@ public class ShopInfo {
 
     private Long rentFee;
 
-    private Integer ratio;
+    private Float ratio;
 
     private Integer rank;
 
@@ -83,11 +83,11 @@ public class ShopInfo {
         this.rentFee = rentFee;
     }
 
-    public Integer getRatio() {
+    public Float getRatio() {
         return ratio;
     }
 
-    public void setRatio(Integer ratio) {
+    public void setRatio(Float ratio) {
         this.ratio = ratio;
     }
